LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/img-src/manifest-v2-img-src-host-only-allowed-expected.txt

ALERT: PASS: Successfully loaded image.

Test that CSP for extensions allows wildcards for other directives.
